fix the pattern can be only run once problem. test=develop
